JobMarket/marketing/requirements.txt
Django==2.2.5
django-rest==0.0.1
djangorestframework==3.10.2
pytz==2019.2
sqlparse==0.3.0
